ft add unit-integration tests

+++ back-etg/src/main/java/etg/example/demo/repository/ITokenRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ITokenRepository extends JpaRepository<Token, Long> {
+public interface ITokenRepository extends JpaRepository<Token, Integer> {
     List<Token> findByTokenStatusAndMeter(boolean tokenStatus, Number meter);
 }
